Remove `initUnlocker` & wrap with our custom iife
The reason we now use a custom iife, is because Rollup would not let us use `return`.
Another reason is to prevent a call to `argument.callee` which
is deprecated and slow. We now have a named iife which we also provide
an argument to determine if our function should be promoted
with `window.eval`.