Babel version (#79)
* chore: Introduce build/bundle process

Using Rollup with additional plugins we bundle our code.

> Used to transpile our code to older browser's feature set, the configuration of browsers can be adjusted in `.browserslistrc`.
For that we use a Babel plugin listed in `babel.config.js`. We also uglify/minify the final output.
> Used to inject our userscript meta, located in `userscript.config.js`

* call rollup without npx

* Let Babel do the minifying & remove Terser

* disable minification

* fix out-of-scope functions

* Support as low as Firefox 68

* removed `@updateURL` & `@downloadURL` properties

* Remove `initUnlocker` & wrap with our custom iife

The reason we now use a custom iife, is because Rollup would not let us use `return`.
Another reason is to prevent a call to `argument.callee` which
is deprecated and slow. We now have a named iife which we also provide
an argument to determine if our function should be promoted
with `window.eval`.

* Fix `import`/`export` not working

* restructuring

* Move `rollup/plugin-node-resolve` to dev dependency

* Refactor notification

* fix

* Move native functions in one file inside utils

* refactor `nativeXMLHttpRequestOpen`

* shuffling around code

* fix notification

* convert `Defered` to class

* Refactor

- Add logger
- Move the inspectYtData logic to main
- Split innertube to innertubeConfig and innertubeClient
And some more small changes...

* remove innertubeConfig, improve error handling

* refactor

* notification: made desktop check more reliable

* Run babel as an output plugin

Running babel as an output plugin has the advantage of formatting the
output without installling additional plugins

* Add `.editorconfig`

* Add eslint

ESlint is now part of the `build` command and will notify you when something is wrong.

Did some research, it's fine to use npx, it will check for local versions first.

* ESlint helped me fix some things

* Fix misplaced comments by Babel

This has a side-effect of adding some whitespaces/newlines in the output

* Small improvements

* small change

* Fix video description & suggested videos

* small refactor

* Add embedded player support

* Add `InnertubeCreator` unlock method and fix #50

* refactor

* Fix strategy iteration

* retrieve player base script from DOM instead ytcfg & small optimizations

* Fix `Strategy 2` on embedded videos and some refactor

* Update README.md

* Move userscript banner to config file

* Create node.js.yml

* Update node.js.yml

* Revert "Update node.js.yml"

This reverts commit ee0b026880460f52ac7d6fb8b578242fc72708f2.

* Revert "Create node.js.yml"

This reverts commit 988f4cd642d846288081a625037619282225d7d4.

* Improve `isUserLoggedIn` logic

* Update package version

Co-authored-by: DrunkWinter <38593134+DrunkWinter@users.noreply.github.com>