[fix] improve extension compatibility
The `Object.defineProperty` override is not needed and only does harm, so remove it and also do some cleanup.

This override is supposed to make our userscript more compatible with other extensions redefining properties that we use, but actually does the opposite. We backup the getter and setter of a prop if it is defined by other extensions, we then replace the getter and setter of the prop with our own. To make it compatible with the other extension we call the getter and setter that we backed up, after we have finished doing things on our side. But then the real issue begins when we override `Object.defineProperty` and intercept the other extensions trying to redefine the prop getter and setter by replacing those with the ones we backed up earlier, thus trading compatibility with the extension that has run before us with the one that ran after.

Obviously that is not what we want. A proper chain starts with us storing the previous getter and setter and calling those when we finish doing our work. The next extension that is trying to define their getter and setter should do the same thing, creating a chain.

A proper chain would look something like this in a callstack.

Before:
- Extension A is detached.
- C is calling B while B is calling C. (potential loop)

A             B - us           C
[get, set]    [get, set] -> <- [get, set]

After:
All extensions call the extension that has ran before them. Creating a proper chain.

A             B - us        C
[get, set] <- [get, set] <- [get, set]