[chore] remove eslint & husky
Currently both are a pain in the ass and hard to work with in terms of developer experience, they are more like a nuisance than a helper.

ESlint is hard to work with when unconfigured. We might add this back in later when we have enough time to write the config rules.

Husky is reasonable, but with Prettier it tends to change files which are not in the commit. Although this is solvable with `lint-staged`, the fact that your code looks different on the commit vs what you just wrote is annoying. Also the error/warning messages are a lap of text, with bunch of useless information and it also does not play nicely with a GUI based Git.

My proposal (#97) as an alternative is having CI testing on GitHub on each pull request, when it fails, the dev must manually run `npm run format`. At that point it is clear to the dev that transformation has happened to the original code and is able to review it.