[chore] replace Prettier with dprint
dprint (written in Rust) is much faster than Prettier by a long shot. Formatting the `user.js` in `dist` took ~400ms with Prettier, now this only takes ~20ms.

I also find the formatting much better with dprint. It reorders all the imports unlike Prettier and some of the multi-line code is more readable.

Out of the box it formats TypeScript, could come in handy if we decide to transition from JavaScript in the future.

This change will impact building, formatting and GitHub CI.