[chore] switch to a build script
This is a more extensible build system and improves performance significantly.

By executing `clean` and `archive-extension` in the same JS context, startup cost of two additional contexts is eliminated.
Benchmark results show an improvement of ~2.7s. Down from ~6.7s to ~4.0s after this patch.

Also by moving away from a config/plugin approach we can easily reorder or insert build steps without too much hassle.

This also lays the groundwork, so we can easily move away from the current bundler if we wanted to.

The logging is temporary to measure performance and for debugging purposes.