[fix] video `startTimeSecs` ignored with proxy (#208)
Account proxy response currently does not include playerConfig.

Until we rewrite the account proxy to only include the necessary and bare minimum response, this will be fixed in the client only. I don't want to take the risk of somehow breaking it as the current proxy codebase IMO is a mess.